clean up docker stuff

+++ docker/entrypoint.sh
@@ -2,24 +2,29 @@
 
 set -e
 
-DJANGO_MOCK=1 DJANGO_POSTGRES=1 python manage.py migrate --noinput
+echo "${@}"
 
-# create users in the database
-DJANGO_MOCK=1 DJANGO_POSTGRES=1 python3 manage.py shell <<-EOF
-	from django.contrib.auth.models import User
-	User.objects.all().delete()
-	User.objects.create_superuser('admin', email='', password='$ADMIN_PASSWORD')
-	User.objects.create_user('mod', password='$MOD_PASSWORD')
-	User.objects.create_user('pad', password='$PAD_PASSWORD')
-EOF
+# initialize the database if the container is run with django related commands
+if [[ "$1" == *"daphne" || "${@}" == *"manage.py"* ]]; then
 
-if [ "$ADMIN_PASSWORD" == "admin" ]; then
-  echo 1>&2
-  echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" 1>&2
-  echo "! Warning! Default admin password used, please consider changing it! !" 1>&2
-  echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" 1>&2
-  echo 1>&2
+	DJANGO_MOCK=1 DJANGO_POSTGRES=1 python manage.py migrate --noinput
+
+	# create users in the database
+	DJANGO_MOCK=1 DJANGO_POSTGRES=1 python3 manage.py shell <<-EOF
+		from django.contrib.auth.models import User
+		User.objects.all().delete()
+		User.objects.create_superuser('admin', email='', password='$ADMIN_PASSWORD')
+		User.objects.create_user('mod', password='$MOD_PASSWORD')
+		User.objects.create_user('pad', password='$PAD_PASSWORD')
+	EOF
+
+	if [ "$ADMIN_PASSWORD" == "admin" ]; then
+	  echo 1>&2
+	  echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" 1>&2
+	  echo "! Warning! Default admin password used, please consider changing it! !" 1>&2
+	  echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" 1>&2
+	  echo 1>&2
+	fi
 fi
 
-echo "Starting: ${@}"
 exec "${@}"
